selenium_py/requirements.txt
selenium
webdriver-manager
pyyaml
python-dotenv
loguru